Fixes for windows.

+++ tests/c/test_derived.c
@@ -28,7 +28,7 @@ TEST(derived, anchors) {
   value_t d0 = new_derived_object_anchor(dgStackPointer, 0);
   ASSERT_EQ(0, get_derived_object_anchor_host_offset(d0));
 
-  uint64_t v1 = 1LLU << 31;
+  uint64_t v1 = ((uint64_t) 1) << 31;
   value_t d1 = new_derived_object_anchor(dgStackPointer, v1);
   ASSERT_EQ(v1, get_derived_object_anchor_host_offset(d1));
